fixup! Update test harness (add assertions) #1467
